requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
idna==3.13
Mako==1.3.12
MarkupSafe==3.0.3
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
python-pytun==2.4.1
requests==2.33.1
scapy==2.7.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
python-jose==3.5.0
passlib==1.7.4
pydantic-settings==2.11.0
